third_party/blink/web_tests/http/tests/security/contentSecurityPolicy/script-src-blocked-error-event.html

<!DOCTYPE html>
<html>
<head>
    <title>Script element has error on bad script-src</title>
    <meta http-equiv="Content-Security-Policy" content="script-src 'self' 'unsafe-inline'">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script>
        setup({single_test: true});

        window.result = false;
        function scriptError() {
            assert_equals(result, false, "Value should not be set");
            done();
        }
        function scriptLoad() {
            assert_unreached("Script loaded");
            done();
        }
    </script>
</head>
    <body>
    <script onload="scriptLoad();" onerror="scriptError();" src="http://localhost:8000/security/contentSecurityPolicy/resources/script-set-value.js"></script>
    </body>
</html>
